separate package contents into a common folder
