+++ sdk/identity/identity-common/src/plugins/provider.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import { VSCodeCredentialFinder } from "../credentials/visualStudioCodeCredentialPlugin";
-import { TokenCachePersistenceOptions } from "../msal/nodeFlows/tokenCachePersistenceOptions";
+import { TokenCachePersistenceOptions } from "../tokenCachePersistenceOptions";
 
 /**
  * The type of an Azure Identity plugin, a function accepting a plugin
