+++ sdk/identity/identity-common/src/tokenCachePersistenceOptions
