+++ sdk/identity/identity-common/src/visualStudioCodeCredentialPlugin.ts
